refactor(home): Error message cart empty

close #374

+++ app/Http/Controllers/OrderController.php
@@ -15,11 +15,16 @@ class OrderController extends Controller
 {
     public function show()
     {
-        $carts = Cart::content();
-        $countries = Country::pluck('name', 'id');
-        return view('orders.order')
-            ->with('carts', $carts)
-            ->with('countries', $countries);
+        if (Cart::count() == 0) {
+            $error = "Cart is empty! Please add products.";
+            return redirect()->route('cartError', ['error' => $error]);
+        } else {
+            $carts = Cart::content();
+            $countries = Country::pluck('name', 'id');
+            return view('orders.order')
+                ->with('carts', $carts)
+                ->with('countries', $countries);
+        }
     }
 
     public function store(Request $request)
@@ -34,20 +39,32 @@ class OrderController extends Controller
             'country_id' => 'required|numeric|exists:countries,id',
         ]);
 
-        $order_id = mt_rand();
-        $data = $request->all();
-        $data['id'] = $order_id;
-        unset($data['_token']);
-        if (!Auth::guest()) {
-            $data['user_id'] = Auth::user()->id;
+        if (Cart::count() == 0) {
+            $error = "Cart is empty! Please add products.";
+            return redirect()->route('cartError', ['error' => $error]);
+        } else {
+            $order_id = mt_rand();
+            $data = $request->all();
+            $data['id'] = $order_id;
+            unset($data['_token']);
+            if (!Auth::guest()) {
+                $data['user_id'] = Auth::user()->id;
+            }
+            Order::insert($data);
+            OrderProduct::store($order_id);
+            $subtotal = Cart::subtotal();
+            Cart::destroy();
+            return redirect()->route('orderInformation', [
+                'order_id' => $order_id,
+                'subtotal' => $subtotal]);
         }
-        Order::insert($data);
-        OrderProduct::store($order_id);
-        $subtotal = Cart::subtotal();
-        Cart::destroy();
+    }
+
+    public function orderInformation($order_id, $subtotal)
+    {
         $order = Order::findOrFail($order_id);
         return view('orders.orderInformation')
-            ->with('order', $order)
-            ->with('subtotal', $subtotal);
+                ->with('order', $order)
+                ->with('subtotal', $subtotal);
     }
 }
